Address some reviews
