+++ src/middlewared/debian/control
@@ -161,6 +161,7 @@ Depends: alembic,
          python3-systemd,
          python3-truenas-api-client,
          python3-truenas-verify,
+         python3-usb,
          python3-websocket,
          python3-zeroconf,
          python3-zettarepl,
