+++ src/middlewared/middlewared/plugins/service_/services/all.py
@@ -1,6 +1,7 @@
 from .cifs import CIFSService
 from .docker import DockerService
 from .ftp import FTPService
+from .incus import IncusService
 from .iscsitarget import ISCSITargetService
 from .mdns import MDNSService
 from .netbios import NetBIOSService
@@ -22,7 +23,6 @@ from .pseudo.libvirtd import LibvirtdService, LibvirtGuestService
 from .pseudo.misc import (
     CronService,
     KmipService,
-    IncusService,
     LoaderService,
     HostnameService,
     HttpService,
